+++ src/sources/EMGrank.h
@@ -3,17 +3,17 @@
 
 void EMGrank_core(
 	// IN parameters
-	const double* Pi,
-	const double* Rho,
+	const float* Pi,
+	const float* Rho,
 	int mini,
 	int maxi,
-	const double* X,
-	const double* Y,
-	double tau,
+	const float* X,
+	const float* Y,
+	float tau,
 	const int* rank,
 	// OUT parameters
-	double* phi,
-	double* LLF,
+	float* phi,
+	float* LLF,
 	// additional size parameters
 	int n,
 	int p,
